Move syntactic XCTest scanner to a separate file
The swift-testing scanner is in its own file and the XCTest scanner should be as well, for consistency.

rdar://126529507
